Started custom speech to text streaming but not there yet
