+++ speech_to_text.py
@@ -1,133 +1,34 @@
-"""
-
-Source: https://github.com/davabase/whisper_real_time/blob/master/transcribe_demo.py
-"""
-#! python3.7
-
-import os
-import numpy as np
-import speech_recognition as sr
+import alsaaudio
+import wave
 import whisper
-import torch
 
-from datetime import datetime, timedelta
-from queue import Queue
-from time import sleep
-from sys import platform
+# Configuration
 
+DEVICE = "plughw:CARD=sofhdadsp,DEV=6"
+CHANNELS = 1
+RATE = 16000
+PERIOD_SIZE = 1024
 
-def main():
-
-    model = "base.en" # choices=["tiny", "base", "small", "medium", "large"]
-    energy_threshold = 1000
-    record_timeout = 2
-    phrase_timeout = 3
-    default_microphone = 'sof-hda-dsp' #None #'pulse'
-
-    # The last time a recording was retrieved from the queue.
-    phrase_time = None
-    # Thread safe Queue for passing data from the threaded recording callback.
-    data_queue = Queue()
-    # We use SpeechRecognizer to record our audio because it has a nice feature where it can detect when speech ends.
-    recorder = sr.Recognizer()
-    recorder.energy_threshold = energy_threshold
-    # Definitely do this, dynamic energy compensation lowers the energy threshold dramatically to a point where the SpeechRecognizer never stops recording.
-    recorder.dynamic_energy_threshold = False
+MODEL = "base.en"  # Whisper model: ["tiny", "base", "small", "medium", "large"]
 
 
-    # Prevents permanent application hang and crash by using the wrong Microphone for Linux
-    source = None
-   
-    if not default_microphone or default_microphone == 'list':
-        print("Available microphone devices are: ")
-        for index, name in enumerate(sr.Microphone.list_microphone_names()):
-            print(f"Microphone with name \"{name}\" found")
-        return
-    else:
-        for index, name in enumerate(sr.Microphone.list_microphone_names()):
-            if default_microphone in name:
-                source = sr.Microphone(sample_rate=16000, device_index=index)
-                break
+inp = alsaaudio.PCM(type=alsaaudio.PCM_CAPTURE, mode=alsaaudio.PCM_NONBLOCK,   
+                     device=DEVICE,
+                    channels=CHANNELS,  periodsize=PERIOD_SIZE)
+
+
+# Load Whisper model
+audio_model = whisper.load_model(MODEL)
+
+frames = []
+num_loops = 1000000
+
+print(f"Recording from {DEVICE}")
+
+for _ in range(num_loops):
+    length, data = inp.read()
+    if length:
+        frames.append(data)
         
-    if not source:
-        RED = '\033[31m'
-        RESET = '\033[0m'
-        print(f"{RED}No microphone found!{RESET}")
-        return
-
-    # Load / Download model
-    audio_model = whisper.load_model(model)
-
-    transcription = ['']
-
-    with source:
-        recorder.adjust_for_ambient_noise(source)
-
-    def record_callback(_, audio:sr.AudioData) -> None:
-        """
-        Threaded callback function to receive audio data when recordings finish.
-        audio: An AudioData containing the recorded bytes.
-        """
-        # Grab the raw bytes and push it into the thread safe queue.
-        data = audio.get_raw_data()
-        data_queue.put(data)
-
-    # Create a background thread that will pass us raw audio bytes.
-    # We could do this manually but SpeechRecognizer provides a nice helper.
-    recorder.listen_in_background(source, record_callback, phrase_time_limit=record_timeout)
-
-    # Cue the user that we're ready to go.
-    print("Model loaded.\n")
-
-    while True:
-        try:
-            now = datetime.utcnow()
-            # Pull raw recorded audio from the queue.
-            if not data_queue.empty():
-                phrase_complete = False
-                # If enough time has passed between recordings, consider the phrase complete.
-                # Clear the current working audio buffer to start over with the new data.
-                if phrase_time and now - phrase_time > timedelta(seconds=phrase_timeout):
-                    phrase_complete = True
-                # This is the last time we received new audio data from the queue.
-                phrase_time = now
-                
-                # Combine audio data from queue
-                audio_data = b''.join(data_queue.queue)
-                data_queue.queue.clear()
-                
-                # Convert in-ram buffer to something the model can use directly without needing a temp file.
-                # Convert data from 16 bit wide integers to floating point with a width of 32 bits.
-                # Clamp the audio stream frequency to a PCM wavelength compatible default of 32768hz max.
-                audio_np = np.frombuffer(audio_data, dtype=np.int16).astype(np.float32) / 32768.0
-
-                # Read the transcription.
-                result = audio_model.transcribe(audio_np, fp16=torch.cuda.is_available())
-                text = result['text'].strip()
-
-                # If we detected a pause between recordings, add a new item to our transcription.
-                # Otherwise edit the existing one.
-                if phrase_complete:
-                    transcription.append(text)
-                else:
-                    transcription[-1] = text
-
-                # Clear the console to reprint the updated transcription.
-                os.system('cls' if os.name=='nt' else 'clear')
-                for line in transcription:
-                    print(line)
-                # Flush stdout.
-                print('', end='', flush=True)
-            else:
-                # Infinite loops are bad for processors, must sleep.
-                sleep(0.25)
-        except KeyboardInterrupt:
-            break
-
-    print("\n\nTranscription:")
-    for line in transcription:
-        print(line)
-
-
-if __name__ == "__main__":
-    main()
+result = audio_model.transcribe(frames)
+text = result['text'].strip()
